pyproject.toml
[project]
name = "dfd-lab"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.8.0",
    "librosa>=0.11.0",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.1",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.3.0",
    "seaborn>=0.12.0",
    "timm>=0.9.0",
    "soundfile>=0.13.1",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "zstandard>=0.24.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchaudio = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
